+++ src/Tars.Net.Core/Internal/ClientProxy.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using System.Reflection;
 using AspectCore.Extensions.Reflection;
 using Tars.Net.Attributes;
@@ -18,9 +19,19 @@ namespace Tars.Net.Internal
             return clientFactory.SendAsync(attribute.ServantName, method.Name, outParameters, method.ReturnParameter, isOneway, attribute.Codec, parameters).Result;
         }
 
-        public void SetClientFactory(IRpcClientFactory clientFactory)
+        private void SetClientFactory(IRpcClientFactory clientFactory)
         {
             this.clientFactory = clientFactory;
+        }
+
+        public static object Create(Type type, IRpcClientFactory rpcClientFactory)
+        {
+            MethodInfo mi = typeof(DispatchProxy).GetMethod(nameof(DispatchProxy.Create), new Type[] { });
+            mi = mi.MakeGenericMethod(type, typeof(ClientProxy));
+            var clientProxy = mi.Invoke(null, null);
+
+            ((ClientProxy)clientProxy).SetClientFactory(rpcClientFactory);
+            return clientProxy;
         }
     }
 }
